# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp AI Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface WhatsappStatus {
  connected: boolean;
  /** @nullable */
  phone?: string | null;
  sessionActive: boolean;
  /** @nullable */
  lastConnected?: string | null;
  qrRequired: boolean;
}

export interface QrCode {
  /** @nullable */
  qr?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  status: string;
}

export interface BusinessProfile {
  id: number;
  businessName: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  workingHours?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  deliveryAvailable: boolean;
  /** @nullable */
  paymentMethods?: string | null;
  /** @nullable */
  returnPolicy?: string | null;
  /** @nullable */
  languages?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BusinessProfileInput {
  businessName?: string;
  category?: string;
  description?: string;
  address?: string;
  workingHours?: string;
  contactNumber?: string;
  deliveryAvailable?: boolean;
  paymentMethods?: string;
  returnPolicy?: string;
  languages?: string;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  variants?: string | null;
  inStock: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  category?: string;
  variants?: string;
  inStock?: boolean;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  variants?: string;
  inStock?: boolean;
  imageUrl?: string;
}

export interface Customer {
  id: number;
  phone: string;
  /** @nullable */
  name?: string | null;
  isVip: boolean;
  isBlocked: boolean;
  aiEnabled: boolean;
  /** @nullable */
  notes?: string | null;
  emotionState: string;
  totalMessages: number;
  /** @nullable */
  lastSeen?: string | null;
  createdAt: string;
}

export interface CustomerUpdate {
  name?: string;
  isVip?: boolean;
  isBlocked?: boolean;
  aiEnabled?: boolean;
  notes?: string;
}

export interface Conversation {
  id: number;
  customerId: number;
  customerPhone: string;
  /** @nullable */
  customerName?: string | null;
  status: string;
  aiHandled: boolean;
  humanTakeover: boolean;
  aiPaused: boolean;
  emotionState: string;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount: number;
  isUrgent: boolean;
  updatedAt: string;
  createdAt: string;
}

export interface Message {
  id: number;
  conversationId: number;
  content: string;
  sender: string;
  senderType: string;
  isAiGenerated: boolean;
  /** @nullable */
  emotionDetected?: string | null;
  createdAt: string;
}

export interface ConversationDetail {
  id: number;
  customerId: number;
  customerPhone: string;
  /** @nullable */
  customerName?: string | null;
  status: string;
  aiHandled: boolean;
  humanTakeover: boolean;
  aiPaused: boolean;
  emotionState: string;
  /** @nullable */
  lastMessage?: string | null;
  unreadCount: number;
  isUrgent: boolean;
  messages: Message[];
  updatedAt: string;
  createdAt: string;
}

export interface MessageInput {
  content: string;
}

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  totalPrice: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  customerId: number;
  productId?: number;
  quantity: number;
  totalPrice: number;
  notes?: string;
}

export interface OrderUpdate {
  status?: string;
  notes?: string;
}

export interface AiConfig {
  id: number;
  enabled: boolean;
  tone: string;
  /** @nullable */
  customInstructions?: string | null;
  confidenceThreshold: number;
  escalationSensitivity: string;
  /** @nullable */
  promptPreview?: string | null;
  updatedAt: string;
}

export interface AiConfigInput {
  enabled?: boolean;
  tone?: string;
  customInstructions?: string;
  confidenceThreshold?: number;
  escalationSensitivity?: string;
  promptPreview?: string;
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
  createdAt: string;
}

export interface FaqInput {
  question: string;
  answer: string;
}

export interface FaqUpdate {
  question?: string;
  answer?: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  read: boolean;
  priority: string;
  /** @nullable */
  relatedId?: number | null;
  createdAt: string;
}

export interface AnalyticsSummary {
  totalCustomers: number;
  activeChats: number;
  aiHandledChats: number;
  humanTakeoverChats: number;
  totalMessages: number;
  escalationRate: number;
  avgResponseTime: number;
  satisfactionScore: number;
  aiHandledPercent: number;
  newCustomersToday: number;
  ordersToday: number;
  revenueToday: number;
}

export interface DailyAnalytics {
  date: string;
  totalConversations: number;
  aiHandled: number;
  humanHandled: number;
  escalations: number;
}

export interface TopQuestion {
  question: string;
  count: number;
  category: string;
}

export interface Settings {
  id: number;
  globalAiEnabled: boolean;
  autoHandoverSensitivity: string;
  responseDelaySeconds: number;
  workingHoursStart: string;
  workingHoursEnd: string;
  /** @nullable */
  autoGreetingMessage?: string | null;
  /** @nullable */
  autoAwayMessage?: string | null;
  language: string;
  updatedAt: string;
}

export interface SettingsInput {
  globalAiEnabled?: boolean;
  autoHandoverSensitivity?: string;
  responseDelaySeconds?: number;
  workingHoursStart?: string;
  workingHoursEnd?: string;
  autoGreetingMessage?: string;
  autoAwayMessage?: string;
  language?: string;
}

export type ListProductsParams = {
  search?: string;
  category?: string;
  inStock?: boolean;
};

export type ListCustomersParams = {
  search?: string;
  vip?: boolean;
  blocked?: boolean;
};

export type ListConversationsParams = {
  status?: ListConversationsStatus;
  search?: string;
};

export type ListConversationsStatus =
  (typeof ListConversationsStatus)[keyof typeof ListConversationsStatus];

export const ListConversationsStatus = {
  all: "all",
  ai: "ai",
  human: "human",
  urgent: "urgent",
  unread: "unread",
} as const;

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  customerId?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type ListNotificationsParams = {
  unreadOnly?: boolean;
};