# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp AI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: vip
          in: query
          schema: { type: boolean }
        - name: blocked
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer settings (VIP, blocked, AI enabled, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── Conversations ─────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, confirmed, processing, shipped, delivered, cancelled] }
        - name: customerId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (status, etc)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── AI Training ──────────────────────────────────────────────────
  /ai/config:
    get:
      operationId: getAiConfig
      tags: [ai]
      summary: Get AI training config
      responses:
        "200":
          description: AI config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
    put:
      operationId: updateAiConfig
      tags: [ai]
      summary: Update AI training config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiConfigInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiConfig"
  /ai/faqs:
    get:
      operationId: listFaqs
      tags: [ai]
      summary: List FAQs
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [ai]
      summary: Create FAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /ai/faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [ai]
      summary: Update FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [ai]
      summary: Delete FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Notifications ─────────────────────────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        phone: { type: string }
        name: { type: ["string", "null"] }
        isVip: { type: boolean }
        isBlocked: { type: boolean }
        aiEnabled: { type: boolean }
        notes: { type: ["string", "null"] }
        emotionState: { type: string }
        totalMessages: { type: integer }
        lastSeen: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, phone, isVip, isBlocked, aiEnabled, emotionState, totalMessages, createdAt]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        isVip: { type: boolean }
        isBlocked: { type: boolean }
        aiEnabled: { type: boolean }
        notes: { type: string }
...
    Order:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerPhone: { type: ["string", "null"] }
        customerName: { type: ["string", "null"] }
        productId: { type: ["integer", "null"] }
        productName: { type: ["string", "null"] }
        quantity: { type: integer }
        totalPrice: { type: number }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, customerId, quantity, totalPrice, status, createdAt, updatedAt]
    OrderInput:
      type: object
      required: [customerId, quantity, totalPrice]
      properties:
        customerId: { type: integer }
        productId: { type: integer }
        quantity: { type: integer }
        totalPrice: { type: number }
        notes: { type: string }
    OrderUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: string }
    AiConfig:
      type: object
      properties:
        id: { type: integer }
        enabled: { type: boolean }
        tone: { type: string }
        customInstructions: { type: ["string", "null"] }
        confidenceThreshold: { type: number }
        escalationSensitivity: { type: string }
        promptPreview: { type: ["string", "null"] }
        updatedAt: { type: string }
      required: [id, enabled, tone, confidenceThreshold, escalationSensitivity, updatedAt]
    AiConfigInput:
      type: object
      properties:
        enabled: { type: boolean }
        tone: { type: string }
        customInstructions: { type: string }
        confidenceThreshold: { type: number }
        escalationSensitivity: { type: string }
        promptPreview: { type: string }
    Faq:
      type: object
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string }
        createdAt: { type: string }
      required: [id, question, answer, createdAt]
    FaqInput:
      type: object
      required: [question, answer]
      properties:
        question: { type: string }
        answer: { type: string }
    FaqUpdate:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
    Notification:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        read: { type: boolean }
        priority: { type: string }
        relatedId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, type, title, message, read, priority, createdAt]
...