# lib/api-zod/src/generated/types/aiConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AiConfig {
  id: number;
  enabled: boolean;
  tone: string;
  /** @nullable */
  customInstructions?: string | null;
  confidenceThreshold: number;
  escalationSensitivity: string;
  /** @nullable */
  promptPreview?: string | null;
  updatedAt: string;
}