# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  isVip?: boolean;
  isBlocked?: boolean;
  aiEnabled?: boolean;
  notes?: string;
}