# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp AI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  customerId: number;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  quantity: number;
  totalPrice: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}